+++ pkg/engine/internal/planner/physical/node_merge.go
