+++ pkg/engine/internal/planner/physical/node_pointersscan.go
